* Fixed serious initialization bug in SerialLCD Component  * Removed self(!) as nonkw-argument for queuelengths - BAD idea,    not very visible and leads to strange effects not even really    impairing the system. Tricky bug was '''TRICKY. * Fixed an integer division in the Engine Component * New testmessage for really dark lcd brightness

+++ ANRV/Controls/Engine.py
@@ -43,7 +43,7 @@ class SimpleEngine(Axon.Component.component):
 
     def SetThrust(self, msg):
         if isinstance(msg.arg, float):
-            target = (self.center + (self.delta / 2) * msg.arg)
+            target = int(self.center + (self.delta / 2.0) * msg.arg)
             print "\n\n\n##### ENGINE TARGET: ", target
 
             byte[0] = 0x84
